Validator 2.0, Fortis Pool, Adjudicators, many other items.

+++ ReserveBlockCore/Services/BlockValidatorService.cs
@@ -36,13 +36,18 @@ namespace ReserveBlockCore.Services
                 return result;
             }
 
-            var verifyBlockSig = SignatureService.VerifySignature(block.Validator, block.Hash, block.ValidatorSignature);
-
-            //validates the signature of the validator that crafted the block
-            if(verifyBlockSig != true)
+            if(block.Height != 0)
             {
-                return result;//block rejected due to failed validator signature
+                var verifyBlockSig = SignatureService.VerifySignature(block.Validator, block.Hash, block.ValidatorSignature);
+
+                //validates the signature of the validator that crafted the block
+                if (verifyBlockSig != true)
+                {
+                    return result;//block rejected due to failed validator signature
+                }
             }
+
+            
             //Validates that the block has same chain ref
             if(block.ChainRefId != BlockchainData.ChainRef)
             {
@@ -62,7 +67,8 @@ namespace ReserveBlockCore.Services
                 Transactions = block.Transactions,
                 Validator = block.Validator,
                 ChainRefId = block.ChainRefId,
-                NextValidators = block.NextValidators
+                TotalValidators = block.TotalValidators,
+                ValidatorAnswer = block.ValidatorAnswer
             };
 
             newBlock.Build();
@@ -113,38 +119,44 @@ namespace ReserveBlockCore.Services
                     result = true;
                     BlockchainData.AddBlock(block);//add block to chain.
                     BlockQueueService.UpdateMemBlocks();//update mem blocks
-                    StateData.UpdateTreis(block); 
+                    StateData.UpdateTreis(block);
 
-                    foreach (Transaction transaction in block.Transactions)
+                    var mempool = TransactionData.GetPool();
+
+                    if(mempool != null)
                     {
-                        var mempool = TransactionData.GetPool();
-
-                        var mempoolTx = mempool.FindAll().Where(x => x.Hash == transaction.Hash);
-                        if(mempoolTx.Count() > 0)
+                        foreach (Transaction transaction in block.Transactions)
                         {
-                            mempool.DeleteMany(x => x.Hash == transaction.Hash);
-                        }
 
-                        //Adds receiving TX to wallet
-                        var account = AccountData.GetAccounts().FindOne(x => x.Address == transaction.ToAddress);
-                        if(account != null)
-                        {
-                            AccountData.UpdateLocalBalanceAdd(transaction.ToAddress, transaction.Amount);
-                            var txdata = TransactionData.GetAll();
-                            txdata.Insert(transaction);
-                        }
 
-                        //Adds sent TX to wallet
-                        var fromAccount = AccountData.GetAccounts().FindOne(x => x.Address == transaction.FromAddress);
-                        if(fromAccount != null)
-                        {
-                            var txData = TransactionData.GetAll();
-                            var fromTx = transaction;
-                            fromTx.Amount = transaction.Amount * -1M;
-                            fromTx.Fee = transaction.Fee * -1M;
-                            txData.Insert(fromTx);
+                            var mempoolTx = mempool.FindAll().Where(x => x.Hash == transaction.Hash);
+                            if (mempoolTx.Count() > 0)
+                            {
+                                mempool.DeleteMany(x => x.Hash == transaction.Hash);
+                            }
+
+                            //Adds receiving TX to wallet
+                            var account = AccountData.GetAccounts().FindOne(x => x.Address == transaction.ToAddress);
+                            if (account != null)
+                            {
+                                AccountData.UpdateLocalBalanceAdd(transaction.ToAddress, transaction.Amount);
+                                var txdata = TransactionData.GetAll();
+                                txdata.Insert(transaction);
+                            }
+
+                            //Adds sent TX to wallet
+                            var fromAccount = AccountData.GetAccounts().FindOne(x => x.Address == transaction.FromAddress);
+                            if (fromAccount != null)
+                            {
+                                var txData = TransactionData.GetAll();
+                                var fromTx = transaction;
+                                fromTx.Amount = transaction.Amount * -1M;
+                                fromTx.Fee = transaction.Fee * -1M;
+                                txData.Insert(fromTx);
+                            }
                         }
                     }
+
                 }
 
                 return result;//block accepted
@@ -157,11 +169,113 @@ namespace ReserveBlockCore.Services
                 StateData.UpdateTreis(block);
                 return result;
             }
-            //Need to add validator validation method.
+
             
         }
 
+        public static async Task<bool> ValidateBlockForTask(Block block, bool blockDownloads = false)
+        {
+            bool result = false;
 
+            var badBlocks = BadBlocksUtility.GetBadBlocks();
+
+            if (badBlocks.ContainsKey(block.Height))
+            {
+                var badBlockHash = badBlocks[block.Height];
+                if (badBlockHash == block.Hash)
+                {
+                    return result;//reject because its on our bad block list
+                }
+            }
+
+            if (block == null) return result; //null block submitted. reject 
+
+            if (block.Height != 0)
+            {
+                var verifyBlockSig = SignatureService.VerifySignature(block.Validator, block.Hash, block.ValidatorSignature);
+
+                //validates the signature of the validator that crafted the block
+                if (verifyBlockSig != true)
+                {
+                    return result;//block rejected due to failed validator signature
+                }
+            }
+
+
+            //Validates that the block has same chain ref
+            if (block.ChainRefId != BlockchainData.ChainRef)
+            {
+                return result;//block rejected due to chainref difference
+            }
+
+            var blockVersion = BlockVersionUtility.GetBlockVersion(block.Height);
+
+            if (block.Version != blockVersion)
+            {
+                return result;
+            }
+
+            var newBlock = new Block
+            {
+                Height = block.Height,
+                Timestamp = block.Timestamp,
+                Transactions = block.Transactions,
+                Validator = block.Validator,
+                ChainRefId = block.ChainRefId,
+                TotalValidators = block.TotalValidators,
+                ValidatorAnswer = block.ValidatorAnswer
+            };
+
+            newBlock.Build();
+
+            //This will also check that the prev hash matches too
+            if (!newBlock.Hash.Equals(block.Hash))
+            {
+                return result;//block rejected
+            }
+
+            if (!newBlock.MerkleRoot.Equals(block.MerkleRoot))
+            {
+                return result;//block rejected
+            }
+                var blockCoinBaseResult = BlockchainData.ValidateBlock(block); //this checks the coinbase tx
+
+                //Need to check here the prev hash if it is correct!
+
+                if (blockCoinBaseResult == false)
+                    return result;//block rejected
+
+                if (block.Transactions.Count() > 0)
+                {
+                    //validate transactions.
+                    bool rejectBlock = false;
+                    foreach (Transaction transaction in block.Transactions)
+                    {
+                        if (transaction.FromAddress != "Coinbase_TrxFees" && transaction.FromAddress != "Coinbase_BlkRwd")
+                        {
+                            var txResult = await TransactionValidatorService.VerifyTX(transaction, blockDownloads);
+                            rejectBlock = txResult == false ? rejectBlock = true : false;
+                        }
+                        else
+                        {
+                            //do nothing as its the coinbase fee
+                        }
+
+                        if (rejectBlock)
+                            break;
+                    }
+                    if (rejectBlock)
+                        return result;//block rejected due to bad transaction(s)
+
+
+                    result = true;
+
+                }
+
+                return result;//block accepted
+            
+
+        }
 
     }
 }
